Separate SensationControls
- Fix voting by separating the component
- Use thinner close button
- Rollback API url
- Update todo's
