+++ src/api/SensumApi.js
@@ -2,8 +2,8 @@
 
 /* Config */
 const config = {};
-// config.serverRoute  = 'https://sensum-server.herokuapp.com/api';
-config.serverRoute = 'http://192.168.0.202:3000/api';
+config.serverRoute  = 'https://sensum-server.herokuapp.com/api';
+//config.serverRoute = 'http://192.168.0.202:3000/api';
 
 /* Auxiliar methods */
 const parseResponse = function(response) {
